*** empty log message ***
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@304 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/common/Logger.h
@@ -18,8 +18,8 @@ class Logger {
   LogType *type;
 
   timepair_t start;
-  double last_logged;
-  double interval;
+  int last_logged;
+  int interval;
   int wrote_header;
 
   string filename;
